SpreadsheetEnginePatchFunction was SpreadsheetEnginePatch
- also renamed sub-classes.
